# Week1/Day4/Assignment-Fastapi/DSUtracker_openapi_postman Collection.json
{
   "openapi":"3.1.0",
   "info":{
      "title":"SE DSU tracker API",
      "version":"1.0.0"
   },
   "paths":{
      "/":{
         "get":{
            "summary":"Root",
            "operationId":"root__get",
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           
                        }
                     }
                  }
               }
            }
         }
      },
      "/standup":{
         "post":{
            "summary":"Create Standup",
            "operationId":"create_standup_standup_post",
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/StandupCreate"
                     }
                  }
               }
            },
            "responses":{
               "201":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get":{
            "summary":"Get Standups",
            "operationId":"get_standups_standup_get",
            "parameters":[
               {
                  "name":"date",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"null"
                        }
                     ],
                     "description":"Filter by date YYYY-MM-DD",
                     "title":"Date"
                  },
                  "description":"Filter by date YYYY-MM-DD"
               },
               {
                  "name":"status",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"null"
                        }
                     ],
                     "description":"Filter by status: active | done",
                     "title":"Status"
                  },
                  "description":"Filter by status: active | done"
               },
               {
                  "name":"name",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"null"
                        }
                     ],
                     "description":"Filter by member name",
                     "title":"Name"
                  },
                  "description":"Filter by member name"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/standup/{name}":{
         "get":{
            "summary":"Get By Name",
            "operationId":"get_by_name_standup__name__get",
            "parameters":[
               {
                  "name":"name",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Name"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/standup/{id}":{
         "put":{
            "summary":"Update Standup",
            "operationId":"update_standup_standup__id__put",
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Id"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/StandupUpdate"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete":{
            "summary":"Delete Standup",
            "operationId":"delete_standup_standup__id__delete",
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Id"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/standup/summary/today":{
         "get":{
            "summary":"Today Summary",
            "operationId":"today_summary_standup_summary_today_get",
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           
                        }
                     }
                  }
               }
            }
         }
      }
   },
   "components":{
      "schemas":{
         "HTTPValidationError":{
            "properties":{
               "detail":{
                  "items":{
                     "$ref":"#/components/schemas/ValidationError"
                  },
                  "type":"array",
                  "title":"Detail"
               }
            },
            "type":"object",
            "title":"HTTPValidationError"
         },
         "StandupCreate":{
            "properties":{
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "did":{
                  "type":"string",
                  "title":"Did"
               },
               "doing":{
                  "type":"string",
                  "title":"Doing"
               },
               "blockers":{
                  "type":"string",
                  "title":"Blockers",
                  "default":"None"
               }
            },
            "type":"object",
            "required":[
               "name",
               "did",
               "doing"
            ],
            "title":"StandupCreate"
         },
         "StandupUpdate":{
            "properties":{
               "did":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Did"
               },
               "doing":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Doing"
               },
               "blockers":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Blockers"
               },
               "status":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Status"
               }
            },
            "type":"object",
            "title":"StandupUpdate"
         },
         "ValidationError":{
            "properties":{
               "loc":{
                  "items":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"integer"
                        }
                     ]
                  },
                  "type":"array",
                  "title":"Location"
               },
               "msg":{
                  "type":"string",
                  "title":"Message"
               },
               "type":{
                  "type":"string",
                  "title":"Error Type"
               },
               "input":{
                  "title":"Input"
               },
               "ctx":{
                  "type":"object",
                  "title":"Context"
               }
            },
            "type":"object",
            "required":[
               "loc",
               "msg",
               "type"
            ],
            "title":"ValidationError"
         }
      }
   }
}